# blog-front/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Sakura Blog API
  description: API documentation for Sakura Blog frontend-backend integration.
  version: 1.0.0
servers:
  - url: /api
    description: Mock Server

paths:
  /auth/login:
    post:
      summary: User Login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  example: 123456
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  /auth/logout:
    post:
      summary: User Logout
      tags: [Auth]
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /auth/me:
    get:
      summary: Get Current User Info
      tags: [Auth]
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'

  /articles:
    get:
      summary: Get Article List
      tags: [Article]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
        - name: keyword
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          list:
                            type: array
                            items:
                              $ref: '#/components/schemas/ArticleSummary'
                          total:
                            type: integer

  /articles/{id}:
    get:
      summary: Get Article Detail
      tags: [Article]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ArticleDetail'

  /articles/{id}/comments:
    get:
      summary: Get Article Comments
      tags: [Article]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Comment'

components:
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: Success

    AuthResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                token:
                  type: string
                user:
                  $ref: '#/components/schemas/User'

    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        avatar:
          type: string
        roles:
          type: array
          items:
            type: string
        bio:
          type: string

    ArticleSummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        summary:
          type: string
        cover:
          type: string
        publishDate:
          type: string
          format: date-time
        views:
          type: integer
        likes:
          type: integer
        comments:
          type: integer
        tags:
          type: array
          items:
            type: string
        author:
          $ref: '#/components/schemas/User'

    ArticleDetail:
      allOf:
        - $ref: '#/components/schemas/ArticleSummary'
        - type: object
          properties:
            content:
              type: string
            favorites:
              type: integer
            isLiked:
              type: boolean
            isFavorited:
              type: boolean

    Comment:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        createdAt:
          type: string
        user:
          $ref: '#/components/schemas/User'
        likes:
          type: integer
        replies:
          type: array
          items:
            $ref: '#/components/schemas/Comment'